requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
Markdown==2.6.7
markdown2==2.3.1
Pillow==3.1.0
